% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anolis.R
\docType{data}
\name{anolis}
\alias{anolis}
\title{Anole data}
\format{
An object of class \code{list} of length 2.
}
\usage{
data(anolis)
}
\description{
Anole data for treedata.table functions. Many
of the traits (e.g. awesomeness, hostility) in this dataset
retrieved from treeplyr are based on random numbers.
}
\author{
Luke Harmon
}
\keyword{datasets}
