% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.treedata.table.R
\name{as.treedata.table}
\alias{as.treedata.table}
\title{Combine tree and data.frame into a single treedata.table object}
\usage{
as.treedata.table(tree, data, name_column = "detect")
}
\arguments{
\item{tree}{A tree of class \code{phylo}}

\item{data}{A dataset in format \code{data.frame}}
}
\value{
treedata.table An object of type \code{treedata.table} containing the tree and data.table
}
\description{
This function takes as input a tree of class \code{phylo} and a \code{data.frame}
and combines them into a treedata.table. A \code{treedata.table} object is
sorted such that the rows in the data.table are matched to the tip.labels
of the phylogeny. Tip.labels on the tree must match a column of tip
names in the input data.frame. The output of this function will be a
treedata.table, which can be manipulated as a data.table.
}
\examples{
data(anolis)
td <- as.treedata.table(tree=anolis$phy, data=anolis$dat)
}
