% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdt_methods.R
\name{print.treedata.table}
\alias{print.treedata.table}
\title{Print method treedata.table objects}
\usage{
\method{print}{treedata.table}(x, ...)
}
\arguments{
\item{x}{an object of class "treedata.table"}

\item{...}{additional arguments passed to "head.treedata.table"}
}
\value{
Function uses prints the tree and the first lines of the
data.table object.
}
\description{
Print method treedata.table objects
}
