% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdt_methods.R
\name{summary.treedata.table}
\alias{summary.treedata.table}
\title{Summarizing treedata.table objects}
\usage{
\method{summary}{treedata.table}(object, ...)
}
\arguments{
\item{object}{an object of class "treedata.table"}

\item{...}{additional arguments passed to "head.treedata.table"}
}
\value{
Function tries to be smart about summarizing the data
and detecting continuous vs. discrete data, as well as whether
any data have missing data. Also returns the taxa that are
dropped from either the original tree or the original data.
}
\description{
Summarizing treedata.table objects
}
\examples{
data(anolis)
td <- as.treedata.table(anolis$phy, anolis$dat)
summary(td)
}
