% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdt_methods.R
\name{tail.treedata.table}
\alias{tail.treedata.table}
\title{Return the last part of an treedata.table object}
\usage{
\method{tail}{treedata.table}(x, ...)
}
\arguments{
\item{x}{a treedata.table object}

\item{...}{Additional arguments passed to head.data.table}
}
\value{
Last part of an treedata.table object
}
\description{
Return the last part of an treedata.table object
}
\examples{
data(anolis)
td <- as.treedata.table(anolis$phy, anolis$dat)
tail(td)
}
