% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdt.R
\name{tdt}
\alias{tdt}
\title{Run a function on a \code{treedata.table} object}
\usage{
tdt(tdObject, ...)
}
\arguments{
\item{tdObject}{A treedata.table object}

\item{...}{A function call.}
}
\value{
Function output
}
\description{
Run a function on a \code{treedata.table} object
}
\details{
This function allows R functions that use trees and data to be run on
\code{treedata.table} objects.
}
\examples{
data(anolis)
td <- as.treedata.table(anolis$phy, anolis$dat)
\dontrun{
tdt(td, geiger::fitContinuous(phy, extractVector(td, SVL), model="BM", ncores=1))
tdt(td, phytools::phenogram(phy, extractVector(td, SVL), quiet=TRUE, spread.labels=FALSE))
}
}
