% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectCharacterType.R
\name{filterMatrix}
\alias{filterMatrix}
\title{Filter a character matrix, returning either all continuous or all discrete characters}
\usage{
filterMatrix(mat, returnType = "discrete")
}
\arguments{
\item{mat}{A character matrix of class data.frame}

\item{returnType}{Either discrete or continuous}
}
\value{
data.frame with only discrete (default) or continuous characters
}
\description{
This function filters a character matrix based on continuous
(e.g., with values 2.45, 9.35, and so on) or discrete characters
(e.g., with values blue, red, yellow).
}
\examples{
data(anolis)
filterMatrix(anolis$dat, "discrete")
}
