% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdt_methods.R
\name{head.treedata.table}
\alias{head.treedata.table}
\title{Return the first part of an treedata.table object}
\usage{
\method{head}{treedata.table}(x, ...)
}
\arguments{
\item{x}{a treedata.table object}

\item{...}{Additional arguments passed to head.data.table}
}
\description{
Return the first part of an treedata.table object
}
\examples{
data(anolis)
td <- as.treedata.table(anolis$phy, anolis$dat)
head(td)
}
