% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treedata.table.R
\name{[.treedata.table}
\alias{[.treedata.table}
\title{Function for perfoming data.table operations on an object of class \code{treedata.table}}
\usage{
\method{[}{treedata.table}(x, ...)
}
\arguments{
\item{x}{An object of class \code{treedata.table}}

\item{...}{Arguments in the structure of \code{data.table} used to perform changes on the \code{treedata.table} object}
}
\value{
A new object of class \code{treedata.table} with \code{$dat} and \code{$phy} corresponding with the changes set to \code{$dat} using (\code{\link{data.table}})'s structure.
}
\description{
This function can be used to subset rows, select and compute on columns (\code{\link{data.table}}.
}
\examples{
data(anolis)
td <- as.treedata.table(anolis$phy, anolis$dat)
td[,SVL]
td[island == "Cuba" & ecomorph == "TG", .(ecomorph, island, SVL)]
td[, head(.SD, 1), by = .(ecomorph, island)]
}
\seealso{
\code{\link{data.table}}
}
