% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treedata.table.R
\name{[[.treedata.table}
\alias{[[.treedata.table}
\title{Function for extract a named vector from an object of class \code{treedata.table}}
\usage{
\method{[[}{treedata.table}(x, ..., exact = TRUE)
}
\arguments{
\item{...}{Column name in class \code{character}}

\item{.x}{An object of class \code{treedata.table}}
}
\value{
A new object of class \code{vector} with names set to labels corresponding to tip labels in the provided \code{treedata.table} object.
}
\description{
This function extracts a named vector for any  trait from an object of class \code{treedata.table}.
}
\examples{
data(anolis)
td <- as.treedata.table(anolis$phy, anolis$dat)
td[["SVL"]]
}
\seealso{
\code{\link{data.table}}
}
